Move WorldState into separate file
